Consoldiated ScriptBuilder code
