+++ ofrak_core/ofrak/service/script_builder.py
@@ -1,11 +1,23 @@
 from collections import OrderedDict
 from dataclasses import dataclass
 from enum import IntEnum
-from typing import Dict, Optional
+import re
+from typing import Dict, List, Optional, Tuple
+from ofrak.model.resource_model import ResourceIndexedAttribute
+from ofrak.core.filesystem import FilesystemEntry
+from ofrak.model.resource_model import Data
+from ofrak_core.ofrak.model.resource_model import ResourceAttributes
+from ofrak_core.ofrak.service.resource_service_i import ResourceAttributeValueFilter, ResourceFilter
 
 from ofrak_type.error import NotFoundError
 from ofrak.resource import Resource
 
+
+class SelectableAttributesError(Exception):
+    """
+    Prompt the user for an attribute to select with
+    """
+    pass
 
 class ActionType(IntEnum):
     UNPACK = 0
@@ -41,7 +53,84 @@ class ScriptBuilder:
 
     def __init__(self):
         self.script_sessions: Dict[bytes, ScriptSession] = {}
-
+        self.var_names: Dict[bytes, str] = {}
+        self.lines: List[str] = []
+        self.selectable_indexes: List[ResourceIndexedAttribute] = [
+            FilesystemEntry.Name,
+            Data.Offset
+        ]  
+        
+    async def _get_selector(self, resource: Resource) -> str:
+        for ancestor in await resource.get_ancestors():
+            if ancestor.get_id() in self.var_names:
+                break
+        attribute, attribute_value = await self._get_selectable_attribute(resource)
+        result = await ancestor.get_children(
+            r_filter=ResourceFilter(
+                tags=resource.get_most_specific_tags(),
+                attribute_filters=[
+                    ResourceAttributeValueFilter(
+                        attribute=attribute, 
+                        value=attribute_value
+                        )
+                ]   
+            )
+        )
+        if len(list(result)) != 1:
+            raise SelectableAttributesError(f"Resource with ID {resource.get_id()} does not have a selectable attribute.")
+        if isinstance(attribute_value, str) or isinstance(attribute_value, bytes):
+            attribute_value = f"\"{attribute_value}\""
+        return f"""await {self.var_names[ancestor.get_id()]}.get_children(
+            r_filter=ResourceFilter(
+                tags={resource.get_most_specific_tags()},
+                attribute_filters=[
+                    ResourceAttributeValueFilter(
+                        attribute={attribute.__name__}, 
+                        value={attribute_value}
+                    )
+                ]   
+            )
+        )
+        """
+    
+    async def _get_selectable_attribute(self, resource:Resource) -> Tuple[ResourceIndexedAttribute, any]:
+        for attribute in self.selectable_indexes:
+            try:
+                await resource.analyze(attribute.attributes_owner)
+                attribute_value = attribute.get_value(resource.get_model())
+            except Exception as e:
+                print(e)
+                continue
+            return attribute, attribute_value
+        raise SelectableAttributesError(f"Resource with ID {resource.get_id()} does not have a selectable attribute.")
+    
+    async def _generate_name(self, resource: Resource) -> str:
+        # Find the most specific tag and use that with a number
+        most_specific_tag = list(resource.get_most_specific_tags())[0].__name__.lower()
+        _, selectable_attribute_value = await self._get_selectable_attribute(resource)
+        name = f"{most_specific_tag}_{selectable_attribute_value}"
+        name = re.sub("[-./\]", "_", name)
+        if name in self.var_names.values():
+            parent = await resource.get_parent()
+            return f"{self.var_names[parent.get_id()]}_{name}"
+        return name
+    
+    async def get_name(self, resource: Resource) -> bytes:
+        if resource.get_id() in self.var_names:
+            return self.var_names[resource.get_id()]
+        if len(list(await resource.get_ancestors())) == 0:
+            self.var_names[resource.get_id()] = "root_resource"
+            self.lines.append("root_resource = await context.create_root_resource_from_file()\n")
+            return "root_resource"
+        parent = await resource.get_parent()
+        if parent.get_id() not in self.var_names:
+            self.get_name(parent)
+        selector = await self._get_selector(resource)
+        name = await self._generate_name(resource)
+        self.lines.append(f"{name} = {selector}")
+        self.var_names[resource.get_id()] = name
+        return name
+    
     async def add_action(
         self,
         resource: Resource,
